+++ R/MazamaCoreUtils.R
@@ -13,7 +13,9 @@
 #' \item{ python style logging }
 #' \item{ simple error messaging }
 #' \item{ cache management }
-#' \item{ date parsing }
+#' \item{ API key handling }
+#' \item{ date parsing and formatting }
+#' \item{ lat/lon validation and uniqueID creation }
 #' \item{ source code linting }
 #' }
 NULL