+++ R/utils-flog.logger.R
@@ -1,11 +1,11 @@
-#' @title Manage loggers
+#' @title Manage appenders for loggers
 #'
-#' @description See \code{futile.logger::\link[futile.logger]{flog.logger}} for details.
+#' @description See \code{futile.logger::\link[futile.logger]{flog.appender}} for details.
 #'
-#' @name flog.logger
+#' @name flog.appender
 #' @keywords internal
 #' @export
-#' @importFrom futile.logger flog.logger
+#' @importFrom futile.logger flog.appender
 NULL
 
 #' @title Manage logger layouts
@@ -17,3 +17,14 @@ NULL
 #' @export
 #' @importFrom futile.logger flog.layout
 NULL
+
+#' @title Manage loggers
+#'
+#' @description See \code{futile.logger::\link[futile.logger]{flog.logger}} for details.
+#'
+#' @name flog.logger
+#' @keywords internal
+#' @export
+#' @importFrom futile.logger flog.logger
+NULL
+