+++ tests/testthat/test-initializeLogging.R
@@ -30,3 +30,31 @@ test_that("initializeLogging() creates files", {
   )
 
 })
+
+test_that("filePrefix is used", {
+
+  logDir <- tempdir()
+  initializeLogging(logDir, filePrefix = "bop_")
+  logger.trace("TRACE message")
+  logger.debug("DEBUG message")
+  logger.info("INFO message")
+  logger.error("ERROR message")
+
+  expect_true(
+    all(file.exists(
+      file.path(logDir, "bop_TRACE.log"),
+      file.path(logDir, "bop_DEBUG.log"),
+      file.path(logDir, "bop_INFO.log"),
+      file.path(logDir, "bop_ERROR.log")
+    ))
+  )
+
+  file.remove(
+    file.path(logDir, "bop_TRACE.log"),
+    file.path(logDir, "bop_DEBUG.log"),
+    file.path(logDir, "bop_INFO.log"),
+    file.path(logDir, "bop_ERROR.log")
+  )
+
+})
+